rag-backend/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
psycopg2-binary
langchain
langchain-google-genai
sentence-transformers
python-dotenv
pydantic
pyyaml
sqlalchemy
